requirement.txt
pytorch-ignite
tqdm
